🚚 chore: cadastro de parceiros
Atividade não finalizada
