+++ front_oracle/src/components/cadastro_parc/index.js
@@ -1,0 +1,3 @@
+import { Titulo } from "./titulo_cad";
+
+export { Titulo };
