+++ front_oracle/src/components/cadastro_parc/titulo_cad/index.js
@@ -1,0 +1,17 @@
+import React from "react";
+import { View, Text } from "react-native";
+import stylesTitulo from "./titulo.styles";
+
+
+export function Titulo() {
+    
+    return (
+        <View>
+          <View style={stylesTitulo.containerTitle}>
+                <Text style={stylesTitulo.texto}>
+                    Cadastro de Parceiros
+                </Text>
+            </View>
+        </View>
+    );
+}
