+++ front_oracle/src/components/index.js
@@ -2,5 +2,6 @@ import { Cabecalho } from "./common/cabecalho/cabecalho";
 import { Buscar } from "./listagem_users/buscar/buscar";
 import { Filtro } from "./common/filtro/filtro";
 import { ListaUsers } from "./listagem_users/lista/listaUsers";
+import { Titulo } from "./cadastro_parc/titulo_cad/index";
 
-export { Cabecalho, Buscar, Filtro, ListaUsers };
+export { Cabecalho, Buscar, Filtro, ListaUsers, Titulo };
