+++ front_oracle/src/views/cadastro_parc/cadastro_parc.styles.js
@@ -1,0 +1,25 @@
+import * as Styles from "../../styles/index";
+import { StyleSheet, StatusBar, Dimensions } from "react-native";
+
+const windowHeight = Dimensions.get("window").height;
+
+const stylesCadastroParc = StyleSheet.create({
+    background: {
+        flex: 1,
+        backgroundColor: Styles.colors.vermelho,
+    },
+    container: {
+        flex: 1,
+        flexDirection: "column",
+        alignItems: "center",
+        backgroundColor: Styles.colors.brancoFundo,
+        borderTopRightRadius: 40,
+        borderTopLeftRadius: 40,
+        marginTop: 20,
+        minHeight: windowHeight - StatusBar.currentHeight,
+        padding: 20,
+    },
+
+});
+
+export default stylesCadastroParc;
