+++ front_oracle/src/views/cadastro_parc/index.js
@@ -1,9 +1,18 @@
-import { View } from 'react-native';
+import { View, SafeAreaView, ScrollView, Text } from 'react-native';
+import * as Components from "../../components/index"
+import stylesCadastroParc from './cadastro_parc.styles';
+
 
 export function CadastroParc() {
   return (
-    <View>
-      
-    </View>
+    <SafeAreaView style={stylesCadastroParc.background}>
+            <ScrollView>
+                <Components.Cabecalho />
+                <Components.Titulo />
+                <View style={stylesCadastroParc.container}>
+              
+                </View>
+            </ScrollView>
+    </SafeAreaView>
   );
 }
